refactor(assert): rename trait rules (#5)
